requirements.txt
matplotlib>=3.2.2
mypy-extensions>=0.4.3
numpy>=1.18.1
pandas>=0.25.1
pyyaml>=5.1
seaborn>=0.10.1
scikit-learn>=0.23.1
scipy>=1.4.1
torch>=1.7.1
tqdm>=4.41.1
prettytable>=2.0.0
